Orders & Payment from  USA

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { BasketService } from 'src/app/basket/basket.service';
 import { CheckoutService } from '../checkout.service';
@@ -6,17 +6,59 @@ import { ToastrService } from 'ngx-toastr';
 import { Basket } from 'src/app/shared/Models/basket';
 import { Address } from 'src/app/shared/Models/user';
 import { NavigationExtras, Router } from '@angular/router';
+import { Stripe, StripeCardCvcElement, StripeCardExpiryElement, StripeCardNumberElement, loadStripe } from '@stripe/stripe-js';
 
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent {
+export class CheckoutPaymentComponent implements OnInit {
+
   @Input() checkoutForm?: FormGroup;
+  @ViewChild('cardNumber') cardNumberElement?: ElementRef;
+  @ViewChild('cardExpiry') cardExpiryElement?: ElementRef;
+  @ViewChild('cardCvc') cardCvcElement?: ElementRef;
+  stripe: Stripe | null = null;
+  cardNumber?: StripeCardNumberElement;
+  cardExpiry?: StripeCardExpiryElement;
+  cardCvc?: StripeCardCvcElement;
+  cardErrors: any;
+  description?: 'test'
 
   constructor(private basketService: BasketService, private checkoutService: CheckoutService,
       private toastr: ToastrService, private router: Router) {}
+
+  ngOnInit(): void {
+
+    loadStripe('pk_test_51NPdH5SJH9yqJ3XPUlUJfkpNQItVfcRltrymZ8h1hQjarTAYMz18aabk1hGjKZ6lEw22rRVslv28VONvxx0gJv0100H5Xz1gHm').then(stripe => {
+      this.stripe = stripe;
+      const elements = stripe?.elements();
+      if(elements) {
+        this.cardNumber = elements.create('cardNumber');
+        this.cardNumber?.mount(this.cardNumberElement?.nativeElement);
+        this.cardNumber.on('change', event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
+
+        this.cardExpiry = elements.create('cardExpiry');
+        this.cardExpiry?.mount(this.cardExpiryElement?.nativeElement);
+        this.cardExpiry.on('change', event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
+
+        this.cardCvc = elements.create('cardCvc');
+        this.cardCvc?.mount(this.cardCvcElement?.nativeElement);
+        this.cardCvc.on('change', event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
+      }
+    })
+    
+  }
 
   submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
@@ -26,9 +68,22 @@ export class CheckoutPaymentComponent {
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: order => {
         this.toastr.success('Order Created successfully');
-        this.basketService.deleteLocalBasket();
-        const navigationExtras: NavigationExtras = {state: order};
-        this.router.navigate(['checkout/success'], navigationExtras);
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!,
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+            },
+          }
+        },
+        ).then(result => {
+          console.log(result);
+          if(result.paymentIntent) {
+            this.basketService.deleteLocalBasket();
+            const navigationExtras: NavigationExtras = {state: order};
+            this.router.navigate(['checkout/success'], navigationExtras);
+          }
+        })
       }
     })
   }
